+++ HISTORY.rst
@@ -11,6 +11,8 @@ History
   chunk size
 * Added ``iter_smart_pk_ranges`` which is similar to ``iter_smart_chunks`` but
   yields only the start and end primary keys for each chunks, in a tuple.
+* Added prefix methods to ``MySQLCache`` - ``delete_with_prefix``,
+  ``get_with_prefix``, ``keys_with_prefix``
 
 0.2.1 (2015-06-22)
 ------------------